实战项目 - 论坛 - 22、使用MyBatis Generator的分页功能（特别注意查询大字段数据要用*withBlob*的方法）

+++ 3、实战项目/community/src/main/java/life/majiang/community/controller/PublishController.java
@@ -32,7 +32,7 @@ public class PublishController {
     public String doPublish(@RequestParam(value = "title") String title,
                             @RequestParam(value = "description") String description,
                             @RequestParam(value = "tag") String tag,
-                            @RequestParam(value = "id") Long id,
+                            @RequestParam(value = "id") Integer id,
                             HttpServletRequest request,
                             Model model) {
         // 方便页面回显，这样页面已经填了的信息不用再重填一次
@@ -69,7 +69,7 @@ public class PublishController {
         question.setDescription(description);
         question.setGmtCreate(System.currentTimeMillis());
         question.setGmtModified(question.getGmtCreate());
-        question.setCreator(Long.valueOf(user.getId()));
+        question.setCreator(user.getId());
         question.setTag(tag);
 
         questionService.createOrUpdate(question);
